pyproject.toml
[tool.poetry]
name = "gpx-tools"
version = "0.1.0"
description = ""
authors = ["ZUB3C <pitzubarev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gpxpy = "^1.6.2"
srtm-py = "^0.3.7"
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gpx-add-elevations = "gpx_tools.gpx_add_elevations:main"
gpx-min-max-elevation = "gpx_tools.gpx_min_max_elevation:main"
gpx-uphill-downhill = "gpx_tools.gpx_uphill_downhill:main"
gpx-pulse-cost = "gpx_tools.gpx_run_pulse_cost:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.3"
mypy = "^1.8.0"

[tool.ruff]
target-version = "py311"
line-length = 99
exclude = [
    ".git",
    ".venv",
    ".idea",
    ".tests",
    ".cache",
    "build",
    "dist",
    "scripts",
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "N", # pep8-naming
    "S", # flake8-bandit
    "B", # flake8-bugbear
    "G", # flake8-logging-format
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "PD", # pandas-vet
    "PLC", # pylint conventions
    "PLE", # pylint errors
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "YTT", # flake8-2020
    "DTZ", # flake8-datetimez
    "RUF", # ruff-specific rules
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ASYNC", # flake8-async
]
ignore = [
    "PTH123", # `open()` should be replaced by `Path.open()`
    "PD901", # Avoid using the generic variable name df for DataFrames
    "RUF003", # Ambiguous unicode character comment
    "RUF001", # String contains ambiguous `В` (CYRILLIC CAPITAL LETTER VE)
]

[tool.mypy]
strict = true
exclude = [
    ".git",
    ".venv",
    ".idea",
    ".tests",
    ".cache",
    "build",
    "dist",
    "scripts",
]